backend/pyproject.toml
[project]
name = "rudapt-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
    "feedparser>=6.0",
    "anthropic>=0.39",
    "openai>=1.40",
    "arq>=0.26",
    "tenacity>=8.5",
    "email-validator>=2.2",
]

[tool.setuptools.packages.find]
include = ["app*"]
